Project management project delete/update

+++ controllers/api/projects.controller.js
@@ -6,6 +6,7 @@ var config = require('config.json');
 var express = require('express');
 var router = express.Router();
 var projectService = require('services/project.service');
+var userService = require('services/user.service');
 var log = require('../logs/effectLog.controller.js');
 
 // routes
@@ -61,17 +62,25 @@ function getListProjects(req, res) {
 
 function updateEffect(req, res) {
 
-    console.log(req.params);
-    console.log(res);
+    //console.log(req.params);
+
+    //console.log(req.user.sub);
+    //console.log(req.body.authorId);
+    if (req.user.sub !== req.body.authorId) {
+        // can only update own account
+        return res.status(401).send('You can only update your own account');
+    }
+
 
     projectService.update(req.params._id, req.body)
         .then(function () {
-            log.info(req.body.title +'\'s account has been updated');
+            log.info(req.body.title + '\'s project has been updated');
             res.sendStatus(200);
         })
         .catch(function (err) {
             res.status(400).send(err);
         });
+
 }
 
 function deleteEffect(req, res) {
@@ -81,13 +90,31 @@ function deleteEffect(req, res) {
         // can only delete own account
         return res.status(401).send('You can only delete your own effect');
     }*/
+    console.log(req.user.sub);
+    userService.getById(req.user.sub)
+        .then(function (user) {
+            if (user) {
+                console.log(user.role.localeCompare('admin')===0);
 
-    projectService.delete(req.params._id)
-        .then(function () {
-            log.info(req.params._id +'\'s effect has been deleted');
-            res.sendStatus(200);
-        })
-        .catch(function (err) {
-            res.status(400).send(err);
-        });
+                projectService.getById(req.params._id)
+                    .then(function(project) {
+                        if (project) {
+                            if((user.role.localeCompare('admin')===0) || (req.user.sub === project.authorId)) {
+                                projectService.delete(req.params._id)
+                                    .then(function () {
+                                        log.info(req.params._id + '\'s effect has been deleted');
+                                        res.sendStatus(200);
+                                    })
+                                    .catch(function (err) {
+                                        res.status(400).send(err);
+                                    });
+                            }else{
+                                res.status(400).send('You don\'t have the right');
+                            }
+                        }
+                })
+            }}).catch(function (err) {
+        res.status(400).send(err);
+    });
+
 }
